services/text-to-speech/requirements.txt
coqui-tts==0.24.1
gradio
fastapi
uvicorn
aiofiles
